Create BART balloon and BART main screen
The BalloonRatioBuilder.kt file was used to make the ratios for the balloon available for various Box() sizes as the original canvas drawing was done using a 300.dp box. I am 98% sure the Balloon.kt file will need to be changed to make the code more readable but I wanted to save some progress here before continuing on. The buttons are abnormally large because they need to be pressable in the underwater case and easily visible.