Update BartTestScreen via ViewModel and uiState
This update adds a BartViewModel and a BartUiState to allow for the propagation of state from the UiState to the BartTestScreen. balloonReward and totalEarnings (required for BartTestScreen) were abstracted to the BartUiState class. balloonRadius and initialBalloonRadius could not be abstracted because they depend on the BalloonCanvas Box size. balloonPopped was therefore added to the BartUiState class.