Feature - Use ViewModel and UiState to update BartTestScreen (#4)
* Rename files to accommodate new Balloon Model

* Create Balloon model class and BalloonGenerator

The BalloonGenerator class has a method that returns a LinkedList with NUMBER_OF_BALLOONS balloons.

* Update BartTestScreen via ViewModel and uiState

This update adds a BartViewModel and a BartUiState to allow for the propagation of state from the UiState to the BartTestScreen. balloonReward and totalEarnings (required for BartTestScreen) were abstracted to the BartUiState class. balloonRadius and initialBalloonRadius could not be abstracted because they depend on the BalloonCanvas Box size. balloonPopped was therefore added to the BartUiState class.

* Move Balloon into data

* BalloonGenerator instance will only make one list

Update the BalloonGenerator class to have the class expose a LinkedList that gets populated in the method. This way we can ensure that each instance of the class can only generate a single list.

* Update BartUiState values for accuracy

* Update ViewModel and BalloonGenerator

Updates to address comments in PR#4

* ViewModel review complete

* Update ViewModel to end test when balloonList is empty