Update generateChars method, Bugs remain
This update has a (mostly) functioning implementation of the code to generate nBack order. Most of the time it is returning a 1:2 validity ratio (target-foil ratio) but occasionally bugs come up in the two-back and three-back conditions where an additional target is created unintentionally. This bug appears to be in the initial block which places targets and buddies. adjacentChars must also be updated for case where i-n to i+n contains all possible stimuli as this causes an exception.