Create ParticipantManagerScreen and Modify/Add Participant Screen (#6)
* Add navigation to participants activity

* Add UI and make cards selectable

* Fix color issues + simplify code

* Navigation to modify/add screen with parcel

* Remove/simplify unneeded code

* Working version of ParticipantsScreen

Some work still needs to be done to clean up code to make it more readable and also will need to add the database

* Organize code for readability

* Build fullscreen dialog UI

* Combine screens into ParticipantFragment

By combining the screens into the one fragment the UiState and ViewModel can be shared, allowing for the same participantList to be updated via the same viewModel

* Add delete participant functionality

* Make PR requested changes

We now have 2 ViewModels, 2 UiStates and 2 Fragments. SafeArgs is used to pass the screenType to AddEditParticipant Screen.

* Try to add Edit functionality

Still two remaining bugs in this code. TextFields populate with participantInfo but do not change with onValueChange. BottomSheet also does not hide() properly.

* fixes editing particpant fields

* adds hiding of bottom sheet

* Fix bottomSheet bug

---------

Co-authored-by: Tony Kazanjian <tony.kazanjian@gmail.com>