NBackGenerator - Working
* Create NBackItem class and Generator

* Create NBackGenerator and Item

* Update generateChars method, Bugs remain

This update has a (mostly) functioning implementation of the code to generate nBack order. Most of the time it is returning a 1:2 validity ratio (target-foil ratio) but occasionally bugs come up in the two-back and three-back conditions where an additional target is created unintentionally. This bug appears to be in the initial block which places targets and buddies. adjacentChars must also be updated for case where i-n to i+n contains all possible stimuli as this causes an exception.

* Update naming to fit psych conventions

* Update names for precision

* Add lureCounter debugging help

* Working implementation; need fix infinite loop bug

* Simplify duplicated code in if/else

* Mitigate stackoverflow bugs

Added counters for any of the while loops to ensure there is a secondary escape condition if a value cannot be placed that the generateChars function recursively calls on itself again to restart the process of generating chars.

* Extract repeated code