Add N-Back Generator and Screen (#8)
* Create NBackItem class and Generator

* Create NBackGenerator and Item

* Update generateChars method, Bugs remain

This update has a (mostly) functioning implementation of the code to generate nBack order. Most of the time it is returning a 1:2 validity ratio (target-foil ratio) but occasionally bugs come up in the two-back and three-back conditions where an additional target is created unintentionally. This bug appears to be in the initial block which places targets and buddies. adjacentChars must also be updated for case where i-n to i+n contains all possible stimuli as this causes an exception.

* Create first NBackScreen

* Remove unneeded old screens

* Remove prints in Generator

* Add uiState and ViewModel

* Working model of NBack

* Create middle NBack item

* Add click feedback

* Limit screen clickability

Only allow screen to be clicked when stimuli is being presented

* Allow only 1 click per stimuli

* Add cycle through N Values

Add the ability to cycle through n=1 then n=2 then n=3

* Update string resources

* Edit viewModel and hoist injection to fragment

* Add custom dialog

* Fix plurals in dialog

* Update times to real values