Feature DSST Part 1 - Implement Practice Screen (#19)
* Create working screen (UI)

* Create working screen (UI)

* Refactor DsstScreen with Viewmodel + MVI

* Connect Dsst screen to navigation

This also allows the screen to be selected for practice.

* Resolve 'Tally Counter' animation bug

Tally counter animation now plays even if two stimuli are the same back to back.

* Add dialog + logic to DSST

Refactored the Dialog component to be more reusable across the app. Added logic to track DSST numClicks, numCorrect, and numIncorrect. Added logic for DsstScreen to only be active for 1 min (practice); 2 mins (real).